+++ source/main/ipc/getHardwareWalletChannel.js
@@ -1,7 +1,7 @@
 // @flow
 import TransportNodeHid from '@ledgerhq/hw-transport-node-hid';
 import { getDevices } from '@ledgerhq/hw-transport-node-hid-noevents';
-import AppBcc, { utils } from '@The-Blockchain-Company/ledgerjs-hw-app-bcc';
+import AppBcc, { utils } from '@theblockchaincompany/ledgerjs-hw-app-bcc';
 import { BrowserWindow } from 'electron';
 import TrezorConnect, {
   DEVICE_EVENT,