+++ source/renderer/app/utils/hardwareWalletUtils.js
@@ -1,7 +1,7 @@
 // @flow
 import _ from 'lodash';
 import { bech32 } from 'bech32';
-import { utils } from '@The-Blockchain-Company/ledgerjs-hw-app-bcc';
+import { utils } from '@theblockchaincompany/ledgerjs-hw-app-bcc';
 import { HARDENED } from '../config/hardwareWalletsConfig';
 
 // Types